+++ src/br/edu/up/modelos/Tripulante.java
@@ -1,30 +1,30 @@
 package br.edu.up.modelos;
 
 public abstract class Tripulante extends Pessoa{
-    protected int idAeronautica;
-    protected int idMatricula;
+    protected double idAeronautica;
+    protected double idMatricula;
     
     
     public Tripulante(){
 
     }
     
-    public Tripulante(String nome, String rg, int idAeronautica, int idMatricula) {
+    public Tripulante(String nome, String rg, double idAeronautica, double idMatricula) {
         super(nome, rg);
         this.idAeronautica = idAeronautica;
         this.idMatricula = idMatricula;
     }
     
-    public int getIdAeronautica() {
+    public double getIdAeronautica() {
         return idAeronautica;
     }
-    public void setIdAeronautica(int idAeronautica) {
+    public void setIdAeronautica(double idAeronautica) {
         this.idAeronautica = idAeronautica;
     }
-    public int getIdMatricula() {
+    public double getIdMatricula() {
         return idMatricula;
     }
-    public void setIdMatricula(int idMatricula) {
+    public void setIdMatricula(double idMatricula) {
         this.idMatricula = idMatricula;
     }
